Started testing the leveling process ...
- We now have a way to create a fake character and a way to level that
  fake characters skills.
  - This is helpful for testing the growth of skills and balancing them
    - Trainable skills can get a max of .99, 99%
    - Non Trainable skills can get a max of .9975, 99.75%

- Changed migrations to use decimals with 5,4 for decimal formatting.
- Changed the way character class and race modifiers are applied.
- Slight Balancing.

+++ app/Flare/Models/GameRace.php
@@ -34,14 +34,14 @@ class GameRace extends Model
      * @var array
      */
     protected $casts = [
-        'str_mod'      => 'integer',
-        'dur_mod'      => 'integer',
-        'dex_mod'      => 'integer',
-        'chr_mod'      => 'integer',
-        'int_mod'      => 'integer',
-        'accuracy_mod' => 'integer',
-        'dodge_mod'    => 'integer',
-        'deffense_mod' => 'integer',
+        'str_mod'      => 'float',
+        'dur_mod'      => 'float',
+        'dex_mod'      => 'float',
+        'chr_mod'      => 'float',
+        'int_mod'      => 'float',
+        'accuracy_mod' => 'float',
+        'dodge_mod'    => 'float',
+        'deffense_mod' => 'float',
     ];
 
     public static function dataTableSearch($query) {
